add ConnectivityStat module
this will help getting more granular reporting of connectivity status
from kernel.  it is linux/android specific for now with possible
future support for bsd/darwin.

this is very much a prototype; cleaner interface and code will follow.

it adds libnl (ships with android) dependency which is really not
required, although expedited prototyping.