configure.ac: Make the Python option depend on the indexation option.
Mimic what is done already in the CMakeLists.txt file, with one
distinction: do not automagically turn on indexation when Python is
selected; rather, stop and tell the user that it is required for
Python support.  Remove the automatic downgrade to --disable-python
when Cython is not found.

* configure.ac: Enforce dependencies on indexation and Cython when
Python support is requested.